Create Reverse_Linked_List_2.java
Here, initially when head == null(Which means the length of the list is 0) and head.next == null(Which means it only has 1 element) and left==right is checked. In all the 3 cases, no changes are made and the same list is returned. 

In the else case, we first create a dummy node with a value -1 and make the dummy.next = head. We first find out the node behind previous and store it in Left_Prev and starting from left until right, we reverse all the nodes, using prev, current and the temporary node and finally when we are done. Link the left prev.next to prev and dummy1.next to current(Which holds the node after right). Make use of dummy node every time you are uncertain about head

Time Complexity: O(n)
Space Complexity : O{1)