Create Find_The_Duplicate_Number.java
The method used here is:
Since it is a given constraint in an array of size N+1, only elements from 1 to N are contained, so i can traverse the array from 0 to N-1 and make each nums[i] negative.(If any nums[i] is already negative, I can make temporarily make it positive and use it. We go and then check if this nums[i] is positive or not, if it is we just continue but if it is negative, then it was already visited. So this is our duplicate element. 

Time Complexity: O(n)
Space Complexity: O(1)