Create Linked_List_Cycle.java
It makes use of 2 pointers, slow and fast. Both are initialised to head and each time slow is incremented by one node and fast by 2 and it is checked if slow is equal to fast and if it is, return true but if fast becomes null and loop ends, return false.