Create Majority_Element_2.java
The Method 1 and Method 2 are the same as the Majority Element Part 1 but as a part of the most Optimized solution, we can use an extended version of Moore's Algorithm, wherein the meaning of counting all the elements greater than n/3 is, utmost, there could be 2 elements in the array whose count could be more than n/3. So instead of having 1 count and element, have count_1 and element_1 for 1 element and count_2 and element_2 for the 2nd element. The logic is I'm trying to first find the 2 most dominating elements in the array, and once i have a value for element_1 and element_2 , I will use the array traversal again to find the frequency of element_1 and element_2 manually and if any of them is more than array_length/3, I append that element that to the answer.

TC: O(n) + O(n)
SC:O(1)