Create Largest_Consequence_Sequence.java
The Brute Force would be a two loop approach where one traverses the array and other finds if number+i is present linearly and in the worst case, the TC exceeds O(n!) and SC is O(1).

The Better Case Approach for this would be to use a Sorting Approach and then when the nearest elements would be groped, the count would be done by ignoring the similar elements. 
TC:O(n.log n)
SC:O(1)

The Best Approach comes with a constraint, Use a Set to Store the unique elements first which is O(n) and then after that traverse the array again and each time check if nums[i] - 1 is present, it skips. The logic is to find the smallest element in the sequence first and the count is done from that part and the overall frequency is also stored in an another variable and that is eventually returned.

TC:(2n)
SC:O(n)