Create Longest_Substring_Without_Repetition.java
Very Similar to the Question: Finding the Longest Sub-Array with the sum as K. The Brute force for this would be a 2 Loop method and the Optimized approach for this would be a to use a HashMap. The logic is, as you go along traversing the array, maintain i for traversing the array and a pointer to keep a track(Initialise it to 0). Until you dont find a element that is repeating, keep moving ahead, but when you do, update the pointer to the elements last index +1 (Stored in HashMap) and update the max_count variable and move ahead.

TC:O(n)
SC:O(n)