Create Unique_BST.java
Here, the logic is basically suppose if n is 3. Then the nodes we have 1,2 and 3. If the root node is 1, then 2 and 3 is to the right and among that 2 and 3, if 2 is taken as root then 3 is to its right but if 3 is taken to the node, then 2 is always taken as the left node. So f(i) at any point will be equal to f(i-1)[Left Part]*f(n-i)[Right Part]. DP is added to minimize the recursive time approach.