Create Validate_BST.java
Here, the usual method is to check at each point if the root.left.val<root.val and root.val<root.right.val. But when the entire tree is considered, it is important that none of the values under its left node are greater than root and none of its values to the right are smaller than the root. To Keep this, a left boundary and right boundary can be maintained each initialised to Long.MIN_VALUE and Long.MAX_VALUE and eventually updated as you go to left and right.

TC:O(n)
SC:O(1)

The Brute force for this would be to check all the nodes on the left and right for all the nodes. This has to happen for all the nodes present.

TC:O(n^2)
SC:O(1)