Create Recover_BST.java
The logic is if the inorder traversal of BST is found out, the order should look like its sorted out in ascending order. 
Brute force method for this, find the inorder traversal first for the tree and store the elements in a array and then sort this array. Traverse the tree comparing it with the sorted array and at any point if the numbers dont match replace the value accordingly. 
TC:O(n)+O(n log n) + O(n)
SC:O(n)

The optimised approach would be to use inorder traversal without the extra space and store the previous nodes value and each time compare it the value of the root node is more than previous node and if it is not store it as first and middle elements and the next element which violates this will be stored as the last element. if the nodes are not adjacent, then first and last could be swapped and if they are adjacent then last will be null. So the first and middle val values can be swapped.
TC:O(n)
SC:O(1)