Create Min_Operations_to_make_X_and_Y_equal.java
Use Recursion as the brute force approach and use Memozation in DP to optimize that.
TC:O(N)
SC:O(N)
Trace the code to understand better. The temporary variable basically increases the value of x to the nearest multiple of 5 and 11 when x itself is not divisible by 5 or 11.