Create Find_Ceil_BST.java
The first method would be to use a Array or List to store all the nodes values by doing an inorder traversal. This will give rise to a sorted array and on this array, key can be compared and a value larger than or equal to key can be computed.
TC:O(N)
SC:O(N)

The Optimised approch would be to do a recursive traversal and have a value to track down the values greater than key and if it is it is checked against the value. Otherwise Recursively move towards left and right.

TC:O(N)
SC:O(N) (Auxillary Stack Space)