Create Delete_Node_In_BST.java
The Logic is Simple, if you delete a node then all the values on the right will be larger than the values on the left at all points. By making use of this, just check that if the node.left is null if it is, then return node.right and do the vice-versa but if both are not null, then make the right detached first and then on the node,left, find the leaf node towards the extreme right and at that  point, point to the current node's right. The Skewed trees poses as a edge-case. Take care accordingly.

TC:O(N)
SC:O(1)