Create combinations_4.java
This combination has to take care of numbers in a different sequence as well. To tackle that, use a 1D DP and make use of target value as a constraint in the DP to negate that. Test every possible present in the nums for all values of the target. Just make a base case for ==0  and >0.

TC:O(n)
SC:O(n)[Recursive Stack Space] + O(n)(DP)