Create Frog_Jump.java
For Bruteoforce, use the method of recursion. Keep track of the number of steps taken in the previous iteration and check for step+1,step,step-1. Return a value of true/false finally.

To optimise it using the method of DP, use a HashMap with each element being Integer,HashSet<Integer> and each element has the elements of stones[i] and set contains the possible steps that can be taken from stones[i]. Iterate over all the elements, add elements to the set based on whether the value is present in the map or not. if it is, add item-1,item,item+1 if value>0. Otherwise skip. If the final element has HashSet size greater than 0, then it means there is a way, otherwise no.

TC:O(n)
SC:O(n) + O(n)