Create M_Coloring.java
Similar approach like Valid Sudoku since one valid way is enough to determine if the final answer is true or false. It tries all colors from1 to M for every node and if anything holds valid, it takes it and moves to the next node. Maintains a colors array to keep track of the colors.

TC:O(n^m)
SC:O(n) + Auxillary Stack Space