Create Median_In_Row_Wise_Sorted_Matrix.java
The test case to use binary search is that if parameters are sent in increasing order, then the return types are also in increasing order.
The idea is, first initialize low to 0 and high to 10^9. Then, the mid is calculated and the number of elements lesser than and equal to mid is calculated and if this is >= (m*n)/2, then first = mid+1, else last = mid-1. Do this until condition fails and finally, first is the median element(The logic is, there reaches a point, where the median's <= elements is way lesser/greater than count and completely opposite >=). 

TC:O(m*n log(m*n))
SC:O(1)