Create Kth_Element_2_Sorted_Arrays.java
The idea when the end goal is to find the max/min of something, there is always a scope of using Binary Search. Here, adjust the low and high such that they never give anything out of bounds or less than zero and apply the same logic as that of finding the median of 2 sorted arrays, but instead its the kth positions. Use the same 4 parameters, l1, l2, r1 and r2.

TC:O(log (Math.max(m,n))
SC:O(1)