Create Course_Schedule_2.java
So arrange the prerequisites in the form of graph(Adjacency List Preferably) and then initialize an array indegree that tracks the incoming edges for all the nodes and whichever has 0, will go into the stack. Do the same for all the nodes and remove the indegrees one at a time. At the end, if the array storing all the elements that have indegree 0 or becomes 0 at some point, then it means there is no cycle and return the array. Otherwise return an Empty Array

TC:O(n)
SC:O(n)