Create Open_The_Lock.java
The Approach used is a BFS since there are no weights attached and you have to reach the destination node from the start in the least amount of steps possible. Maintain a set for visited and deadend nodes and store the nodes in queue if not in either of them. Each time store the node value and the level number. 

TC:O(n)["0000" -> "9999"]
SC:O(n) + O(n)